Tests: Move conftest.py to PyInstaller/utils.
